#1617 merge dto pattern examples
